Move invite initialization code to the invitation manager.
<h1>Rationale</h1>

This commit provides no functional changes.  It simply moves code from MainActivity to InvitationManager.
